src/pmqc/extracao/requirements.txt
pandas
pandas_gbq
requests
google-cloud-bigquery
google-cloud-storage
dotenv
beautifulsoup4

